:recycle: [refactor] 스테이지 실시간 채팅 ui 위젯으로 분리
